Rename calico package
Conflicts with core's calico package, installed by Felix.
