+++ calico_docker/adapter/docker_restart.py
