+++ calico_docker/adapter/requirements.txt
