+++ calico_docker/calicoctl.py
@@ -60,13 +60,13 @@ from netaddr import IPNetwork, IPAddress
 from netaddr.core import AddrFormatError
 from prettytable import PrettyTable
 
-from calico.adapter.datastore import (ETCD_AUTHORITY_ENV,
+from calico_docker.adapter.datastore import (ETCD_AUTHORITY_ENV,
                                       ETCD_AUTHORITY_DEFAULT,
                                       Rules,
                                       DataStoreError)
-from calico.adapter.docker_restart import REAL_SOCK, POWERSTRIP_SOCK
-from calico.adapter.ipam import IPAMClient
-from calico.adapter import netns, docker_restart
+from calico_docker.adapter.docker_restart import REAL_SOCK, POWERSTRIP_SOCK
+from calico_docker.adapter.ipam import IPAMClient
+from calico_docker.adapter import netns, docker_restart
 from requests.exceptions import ConnectionError
 from urllib3.exceptions import MaxRetryError
