+++ calico_docker/tests/fv/diags.sh
