+++ calico_docker/tests/fv/no_powerstrip.sh
