+++ calico_docker/tests/scale/cloud-config-master.yaml
