+++ calico_docker/tests/scale/cloud-config.yaml
