+++ calico_docker/tests/scale/kill.sh
