+++ run_fvs.sh
@@ -42,13 +42,13 @@ docker exec -t host1 bash -c \
 
 # Run the FVs. Need to run from the /code directory since the tests expect
 # to be run from the root of the codebase.
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/mainline.sh'
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/add_container.sh'
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/add_ip.sh'
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/arg_parsing.sh'
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/profile_commands.sh'
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/no_powerstrip.sh'
-docker exec -t host1 bash -c 'cd /code && sudo ./calico/tests/fv/diags.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/mainline.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/add_container.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/add_ip.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/arg_parsing.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/profile_commands.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/no_powerstrip.sh'
+docker exec -t host1 bash -c 'cd /code && sudo ./calico_docker/tests/fv/diags.sh'
 
 docker exec -t host1 bash -c 'docker rm -f $(docker ps -qa) ; \
                               docker rmi $(docker images -qa)' || true
